Collate all test signal spoofers
This avoids the otherwise depends on previous written test dependency
that was being built up.
