+++ VHF-rs/src/runner/test/v1_write.rs
@@ -3,25 +3,21 @@
 use super::Config;
 use super::consts::{MMAP_PAGE_LEN, VHF_MMAP_WINDOW_LEN};
 use super::fold::{StreamFold, StreamFoldOp};
-use super::vhf_step_fold::SineArr;
+use super::signals::{LinearPhaseArr, SineArr};
 use super::writer::builder::Writers;
 use super::{debug_vhf_new, push_arc_pages};
-use crate::{Error, Result};
 
 use jiff::Zoned;
 use std::collections::HashMap;
 use std::f64::consts::TAU;
 use std::ffi::CString;
 use std::num::NonZeroUsize;
-use std::sync::{
-    Arc,
-    atomic::{AtomicBool, AtomicU64, Ordering},
-};
 use std::time::Duration;
 use tempfile::TempDir;
 use test_log::test;
-use vhf_common::data_types::{Polar, RawVHFWord};
 
+#[cfg(feature = "o3")]
+use crate::{Error, Result};
 #[cfg(feature = "o3")]
 use approx::assert_relative_eq;
 #[cfg(feature = "o3")]
@@ -430,72 +426,6 @@ fn creates_correct_multithreaded_files() {
 
     tmp_dir.close().expect("Could not close temp_dir.");
     push_arc_pages_thread.join().expect("Failed to join");
-}
-
-pub(super) struct LinearPhaseArr {
-    total_len: u64,
-    current_idx: AtomicU64,
-    c: f64,
-    m: f64,
-    radius: f64,
-    engine_running: Arc<AtomicBool>,
-}
-
-impl Iterator for LinearPhaseArr {
-    type Item = RawVHFWord;
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.current_idx.load(Ordering::Acquire) >= self.total_len {
-            self.engine_running.fetch_and(false, Ordering::AcqRel);
-            None
-        } else {
-            let j = self.current_idx.fetch_add(1, Ordering::AcqRel) as f64;
-            let curr_phase = self.m * j + self.c;
-
-            let polar = Polar {
-                radius: self.radius,
-                phase: curr_phase * TAU,
-            };
-
-            Some(polar.into())
-        }
-    }
-}
-
-impl LinearPhaseArr {
-    /// Arguments:
-    ///
-    /// - total_len: Number of elements in iterator.
-    /// - initial_params: (radius, initial_reduced_phase, reduced_gradient)
-    ///   Radius is the value of the signal.
-    ///   Reduced phase translates to Unwrapped phase / TAU.
-    pub(super) fn new(
-        total_len: usize,
-        initial_params: (f64, f64, f64),
-        engine_running: Arc<AtomicBool>,
-    ) -> Self {
-        if total_len % MMAP_PAGE_LEN != 0 {
-            log::warn!("LinearArr did not receive an integer multiple of MMAP_PAGE_LEN");
-        }
-        Self {
-            total_len: total_len.try_into().unwrap(),
-            current_idx: AtomicU64::new(0),
-            radius: initial_params.0,
-            c: initial_params.1,
-            m: initial_params.2,
-            engine_running,
-        }
-    }
-}
-
-impl Clone for LinearPhaseArr {
-    /// XXX: This will detach from the [`engine_running`].
-    fn clone(&self) -> Self {
-        Self {
-            current_idx: AtomicU64::new(self.current_idx.load(Ordering::Acquire)),
-            engine_running: Arc::new(AtomicBool::new(false)),
-            ..*self
-        }
-    }
 }
 
 /// Test the basic case of Python parsing first with Linear, and how that checks out against the
